Initial commit master

+++ Jenkinsfile
@@ -3,11 +3,11 @@ pipeline {
 
     stages {
         stage ('Stop and delete running docker container') {
-            steps {
-                sh 'whoami'
-                sh 'groups jenkins'
-                sh 'docker kill $(docker ps -q)'
-                sh 'docker rm $(docker ps -a -q)'
+            script {
+                if (docker ps -q) {
+                    sh 'docker kill $(docker ps -q)'
+                    sh 'docker rm $(docker ps -a -q)'
+                }
             }
         }
         stage ('Build Docker Image') {
